requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.2
pydantic>=2.0.0
python-multipart>=0.0.6
pillow>=10.0.0
pytesseract>=0.3.10
pandas>=2.0.0
numpy>=1.24.0
transformers>=4.30.0
torch>=2.0.0
openpyxl>=3.1.2